add the table_check module.
