+++ qsequoia2/scripts/forest_data/forest_data.py
@@ -19,7 +19,7 @@ from PyQt5.QtCore import Qt
 from .update_forest_name import *
 from ..utils.messageBar import *
 from .forest_get_data import *
-from .data_table import *
+from ..table_check.data_table import *
 from ..utils.variable import *
 from ..utils.seq_config import *
 
@@ -52,16 +52,6 @@ class ForestDataTabs(QDialog, FORM_CLASS):
             cb.setVisible(False)
             cb.toggled.connect(self.on_checkbox_toggled)
 
-        self.forestTable.setVisible(False)
-
-        # réaction au changement de type de données à afficher
-        self.cb_dataType.currentTextChanged.connect(self.actu_Tabledata)
-        # réaction au changement de parcelle sélectionnée
-        self.cb_parcelle.currentTextChanged.connect(self.on_cb_parcelle_changed)
-        self.cb_sspf.currentTextChanged.connect(self.on_cb_parcelle_changed)
-
-
-
 
     # endregion
     # ================================================
@@ -80,8 +70,7 @@ class ForestDataTabs(QDialog, FORM_CLASS):
 
         for cb in self.forestType_checkbox:
             cb.setVisible(True)
-        self.cb_dataType.setEnabled(True)
-        
+
         # création des métadata 
         seq_metadata = self.run_calculation(seq_dir)
         self.export_to_project_variables(seq_metadata, seq_dir)
@@ -205,156 +194,4 @@ class ForestDataTabs(QDialog, FORM_CLASS):
 
         self.forest_name_edit.setText(str(forest_name))
 
-
     # endregion
-    # ================================================
-    # region tableaux
-    # ================================================
-
-    def actu_Tabledata(self, value):
-        seq_dir = Path(get_project_variable("QS2_seq_dir"))
-
-        if not seq_dir:
-            return
-
-        if value == "Vérificateur de données":
-            self.check_data(seq_dir)
-            self.current_layer = self.ua_layer
-            self.forestTable.setVisible(True)
-            self.cb_parcelle.setEnabled(True)
-            self.cb_parcelle.clear()
-            self.cb_sspf.setEnabled(True)
-            self.cb_sspf.setVisible(True)
-            self.lbl_sspf.setVisible(True)
-            self.populate_cb_from_field(cb_name = "cb_parcelle", Layer = self.ua_layer, field_name = "N_PARFOR")
-            self.populate_cb_from_field(cb_name = "cb_sspf", Layer = self.ua_layer, field_name = "N_SSPARFOR")
-
-        elif value == "Synthèse":
-            try:
-                synthese = seq_read("summary", seq_dir)
-                self.final_layer = self.setFinaldata(seq_dir, synthese)
-                self.current_layer = self.final_layer
-
-                self.populate_cb_from_field(cb_name = "cb_parcelle", Layer = self.final_layer, field_name = "N_PARFOR")
-                self.forestTable.setVisible(True)
-                self.lbl_sspf.setVisible(False)
-                self.cb_sspf.setVisible(False)
-                self.cb_parcelle.setEnabled(True)
-                self.fill_table(self.final_layer, list(self.final_layer.getFeatures()))
-            except Exception as e:
-                messageLog(f"Erreur lors de la mise à jour de la table : {e}", "w")
-                
-        
-        elif value == "Sélectionner une table":
-
-            self.cb_parcelle.setEnabled(False)
-            self.cb_sspf.setEnabled(False)
-            self.forestTable.setVisible(False)
-
-    # création de la synthèse et affichage dans la table
-
-    def setFinaldata(self, seq_dir, synthese):
-        """"""
-        source = synthese.dataProvider().dataSourceUri()
-        synthese = source.split("|")[0]
-
-        final_data = getFinaldata(synthese)
-
-        return final_data
-
-    # ================================================
-    # tableaux : Vérification
-    # ================================================
-
-    def check_data(self, seq_dir):
-
-        self.forestTable.clearContents()
-        self.forestTable.setRowCount(0)
-
-        self.fill_table(self.ua_layer, list(self.ua_layer.getFeatures()))
-
-
-
-    # endregion
-    # ================================================
-    # region utilitaires tableaux
-    # ================================================ 
-    
-    def populate_cb_from_field(self, cb_name, Layer, field_name):
-        # Récupérer les valeurs uniques
-
-        cb = getattr(self, cb_name)
-        values = sorted({str(f[field_name]) for f in Layer.getFeatures()})
-        cb.clear()
-        cb.addItem("Toutes")      # valeur par défaut
-        cb.addItems(values)
-        cb.setCurrentIndex(0)
-
-
-    def on_cb_parcelle_changed(self, value):
-
-        layer = self.current_layer
-        layer.removeSelection()
-
-        # Construction dynamique des filtres
-        filters = []
-        pf_value = self.cb_parcelle.currentText()
-        sspf_value = self.cb_sspf.currentText()
-
-        if value != "Toutes":
-            filters.append(f"\"N_PARFOR\" = '{pf_value}'")
-
-        if "N_SSPARFOR" in [field.name() for field in layer.fields()]:
-            if sspf_value != "Toutes":
-                filters.append(f"\"N_SSPARFOR\" = '{sspf_value}'")
-
-        # Si aucun filtre → tout afficher
-        if not filters:
-            self.update_table_with_all(layer)
-            return
-
-        expr = " AND ".join(filters)
-
-        request = QgsFeatureRequest().setFilterExpression(expr)
-        ids = [f.id() for f in layer.getFeatures(request)]
-
-        layer.selectByIds(ids)
-        self.update_table_with_selection(layer, ids)
-
-
-    def update_table_with_all(self, layer):
-        feats = list(layer.getFeatures())
-        self.fill_table(layer, feats)
-
-    def update_table_with_selection(self, layer, ids):
-        feats = [f for f in layer.getFeatures() if f.id() in ids]
-        self.fill_table(layer, feats)
-    
-
-    def fill_table(self, layer, features):
-        """Remplit la table avec les features données"""
-
-        fields = layer.fields()
-
-        self.forestTable.setColumnCount(len(fields))
-        self.forestTable.setHorizontalHeaderLabels([f.name() for f in fields])
-        self.forestTable.setRowCount(len(features))
-
-        # met en forme la table
-        self.TableFormat()
-
-        for row, feat in enumerate(features):
-            for col, field in enumerate(fields):
-                val = feat[field.name()]
-                item = QTableWidgetItem(str(val))
-                item.setFlags(item.flags() ^ Qt.ItemIsEditable)
-                self.forestTable.setItem(row, col, item)
-
-    def TableFormat(self):
-        """Met en forme la table"""
-        header = self.forestTable.horizontalHeader()
-        header.setSectionResizeMode(0, QHeaderView.ResizeToContents)
-        header.setSectionResizeMode(1, QHeaderView.Stretch)
-        header.setSectionResizeMode(2, QHeaderView.ResizeToContents)
-
-        self.forestTable.setAlternatingRowColors(True)
